requirements.txt
Flask==2.0.1
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scipy==1.6.0
scikit-image==0.18.1
six==1.15.0
tifffile==2021.2.1
imutils==0.5.4
opencv-python==4.5.5.64